% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{closest.date}
\alias{closest.date}
\title{Determine the closest date}
\usage{
closest.date(D, D.list)
}
\arguments{
\item{D}{a date}

\item{D.list}{a vector of dates}
}
\value{
the closest date
}
\description{
Determine the closest date
}
\examples{
closest.date("1991-03-30",D.list=as.Date(runif(10,0,20),origin="1991-03-15"))
}
