% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishing_areas}
\alias{fishing_areas}
\title{Fishing area}
\format{A data frame with 150 rows and 2 variables:
\describe{
  \item{fishing_area}{code of the fishing area}
  \item{description}{description of the fishing area}
}}
\usage{
fishing_areas
}
\description{
A dataset of fishing areas given as DCF level 3 (DCF level 4 for the Baltic).
}
\keyword{datasets}

