% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gear_codes}
\alias{gear_codes}
\title{Gear codes}
\format{A data frame with 61 rows and 3 variables:
\describe{
  \item{gear_code}{code of the gear}
  \item{gear_type}{type of the gear}
  \item{description}{description of the gear}
  \item{passive}{is the gear passive (TRUE) or active (FALSE)}
}}
\usage{
gear_codes
}
\description{
A dataset of gear codes taken from the Master Data Register.
}
\keyword{datasets}
