% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ices_rectangles}
\alias{ices_rectangles}
\title{ICES rectangles}
\format{A data frame with 7068 rows and 5 variables:
\describe{
  \item{ices_rectangle}{code of the rectangle}
  \item{south}{southern coordinate}
  \item{west}{westerly coordinate}
  \item{north}{northern coordinate}
  \item{east}{easterly coordinate}
}}
\usage{
ices_rectangles
}
\description{
A dataset of ICES rectangles.
}
\keyword{datasets}

