% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{keep_with_before.after}
\alias{keep_with_before.after}
\title{Subset lines for which there are data before and after treatment
Internal use}
\usage{
keep_with_before.after(
  df,
  col.to.keep = c("id", "period", "tvnt", "species", "seasonchron"),
  col.to.check = "period",
  bef.aft.names = c("before", "after")
)
}
\arguments{
\item{df}{a data frame to subset}

\item{col.to.keep}{list of column names for which data presence should be checked for before and after}

\item{col.to.check}{name of the colum stating either "before" or "after". Default is "period".}

\item{bef.aft.names}{optional if the variable names of the colum col.to.check aren't "before" and "after"}
}
\value{
a subsetted data frame
}
\description{
Subset lines for which there are data before and after treatment
Internal use
}
\examples{
#internal use
}
