% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_tools.R
\name{rand.ind.list}
\alias{rand.ind.list}
\title{Create a random list with a minimum number of individuals
Internal use.}
\usage{
rand.ind.list(FECR.ind, s, group, per, sp)
}
\arguments{
\item{FECR.ind}{data frame of epg data}

\item{s}{character, season to consider}

\item{group}{character, group to consider (e.g. "control", "treated")}

\item{per}{character, period to consider (e.g. "before", "after")}

\item{sp}{character, parasite species to consider (e.g. "oes", "trich")}
}
\value{
an ordered vector of inedividual names
}
\description{
Create a random list with a minimum number of individuals
Internal use.
}
\examples{
#internal use.
}
