% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_desc.R
\docType{data}
\name{Treatment}
\alias{Treatment}
\title{Data on individual treatment}
\format{
An object of class \code{data.frame} with 258 rows and 7 columns.
}
\usage{
Treatment
}
\description{
Data on individual treatment
}
\references{
\url{github.com/R-KenK}
}
\author{
Kenneth Keuk \email{kenneth.keuk@gmail.com}
}
\keyword{datasets}
