% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_desc.R
\docType{data}
\name{EPG}
\alias{EPG}
\title{Data on individual fecal samples and their EPG}
\format{
An object of class \code{data.frame} with 1932 rows and 31 columns.
}
\usage{
EPG
}
\description{
Data on individual fecal samples and their EPG
}
\references{
\url{github.com/R-KenK}
}
\author{
Kenneth Keuk \email{kenneth.keuk@gmail.com}
}
\keyword{datasets}
