% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FECR.R
\name{FECR}
\alias{FECR}
\title{Fecal egg count reduction
Calculate the FECR according to several methods described (cf. Cabaret & Berrag (2004))}
\usage{
FECR(
  T1 = NULL,
  T2 = NULL,
  C1 = NULL,
  C2 = NULL,
  method = c("Kochapakdee", "Dash", "Coles", "Cabaret1", "Cabaret2")
)
}
\arguments{
\item{T1}{pre-treatment mean epg (or vector of individual epg) of the treated group.}

\item{T2}{post-treatment mean epg (or vector of individual epg) of the treated group.}

\item{C1}{pre-treatment mean epg (or vector of individual epg) of the control group.}

\item{C2}{post-treatment mean epg (or vector of individual epg) of the control group.}

\item{method}{method to base the FECR calculation on:
\itemize{
 \item{"Kochapakdee"}{the ratio of pre- and post-treatment epg arithmetic means of the treated group}
 \item{"Dash"}{the ratio of pre- and post-treatment epg arithmetic means of the treated group, correcting for the evolution of mean epg in the control group.}
 \item{"Coles"}{the ratio of the treated and control post-treatment epg arithmetic means}
 \item{"Cabaret1"}{the mean ratio of individual pre- and post-treatment epg of the treated group}
 \item{"Cabaret2"}{the mean ratio of pre- and post-treatment epg of the treated group, correcting for the individual epg in the control group.}
}}
}
\value{
the fecal egg count reduction, in percentage
}
\description{
Fecal egg count reduction
Calculate the FECR according to several methods described (cf. Cabaret & Berrag (2004))
}
\details{
for more informations, cf. Cabaret & Berrag (2004) and the literature they refer to.
}
\examples{
set.seed(42)
T1<- rpois(50,100);T1.mean<- mean(T1)
T2<- rpois(50,10);T2.mean<- mean(T2)
C1<- rpois(50,100);C1.mean<- mean(C1)
C2<- rpois(50,90);C2.mean<- mean(C2)

FECR(T1.mean,T2.mean,method = "Kochapakdee")
FECR(T1,T2,C1,C2,method = "Dash")
FECR(T2 = T2.mean,C2 = C2.mean,method = "Coles")
FECR(T2 = T2,C2 = C2,method = "Coles")
FECR(T1,T2,method = "Cabaret1")
FECR(T1,T2,C1,C2,method = "Cabaret2")
}
